Create main.py
Other than completing the original goals, I've added a feature that allows the computer send feedback about whether the player is getting close or not so that this game is not only about taking "wild guesses" but to have some strategic fun.
Also, I've added some dialogues so that this simple game is more lively. I originally planned to write some hidden background stories if the player keeps messing around and not playing by the rules (e.g.: entering wrong numbers or entering words).  Everytime the player messes around with the "gamebot" it adds up a "hidden story point". As the points continue to accumulate, the gamebot will get super mad and tell the truth: it turn out that this boring game is a trap  and he, the gamebot, was the previous player who kept messing around and ended up being prisoned by the game. The gamebot's original purpose was to lure the player so that he could replace him to be the next prisoner of the game... With all these in mind I am able to construct the very very rubric of it, a prototype, I should say. If possible, I'd like to cooperate with others to finish this project.